+++ react-ui/src/components/pages/DeleteSupplier.js
@@ -8,14 +8,16 @@ function DeleteSupplier() {
     const [supplierId, setSupplierId] = useState();
 
     const handleDelete = (supplierId) =>{
-        console.log("Supplier deleted successfully");
-        SupplierService.deleteSupplier(supplierId)
-        .then(response =>{
-            console.log("Supplier deleted successfully")
-        })
-        .catch(error =>{
-            console.log("Error", error)
-        })
+        if (window.confirm("Do you want to delete supplier id: " + supplierId)) {
+            SupplierService.deleteSupplier(supplierId)
+                .then(response => {
+                    alert("Supplier deleted successfully")
+                    console.log("Supplier deleted successfully")
+                })
+                .catch(error => {
+                    console.log("Error", error)
+                })
+        }
     }
 
     const handleReset = () => {
@@ -29,7 +31,7 @@ function DeleteSupplier() {
                     <div className="col-md-6">
                         <form className="col-md-4">
                             <label for="supplierId" className="form-label">Enter Supplier ID:</label>
-                            <input type="number" min="1" className="form-control" id="supplierId" />
+                            <input type="number" min="1" className="form-control" id="supplierId"  value={supplierId} onChange={(e) => setSupplierId(e.target.value)} />
                             <div className="mt-3">
                                 <button type="submit" className="btn btn-danger btn-sm search-btns" onClick={() => handleDelete(supplierId)}>Delete</button>
                                 <button type="reset" className="btn btn-secondary btn-sm search-btns" onClick={handleReset}>Reset</button>
